+++ src/pages/Login.js
@@ -33,9 +33,6 @@ const Login = () => {
       .post("http://3.34.188.26/user/login", userLoginData)
       .then(function (response) {
         localStorage.setItem("access_token", response.headers.authorization);
-        localStorage.setItem("userLogin", true);
-        const UserState = localStorage.getItem("UserLogin");
-        dispatch(OnLoginUser(UserState));
         alert("success");
         navigate("/");
         console.log(response);
